testing/pyproject.toml
[project]
name = "commons-testing"
version = "0.1.0"
description = "Testing utilities for Python Commons"
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Python Commons Team", email = "team@pythoncommons.org"}
]
dependencies = [
    "commons-core",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-timeout>=2.1.0",
    "pytest-xdist>=3.3.0",
    "faker>=19.0.0",
    "factory-boy>=3.3.0",
    "hypothesis>=6.80.0",
    "testcontainers>=3.7.0",
    "httpx>=0.24.0",
    "respx>=0.20.0",
]


[tool.uv.sources]
commons-core = { workspace = true }

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
databases = [
    "pytest-postgresql>=5.0.0",
    "pytest-mysql>=2.0.0",
    "pytest-mongodb>=2.4.0",
    "pytest-redis>=3.0.0",
]
cloud = [
    "moto[all]>=4.2.0",
    "localstack>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/commons_testing"]